requirements.txt
pandas>=2.2,<3.0
pypdf>=5.0,<6.0
reportlab>=4.2,<5.0
python-dateutil>=2.9,<3.0
streamlit>=1.45,<2.0
plotly>=5.24,<6.0
